feat(bench): add reproducible benchmark harness
- split benchmark code into focused modules
- add aiohttp comparison client
- render reports with Jinja templates
- switch benchmark CLI to Typer
- document benchmark results
- ignore generated benchmark outputs
