+++ Cargo.toml
@@ -11,9 +11,44 @@ crate-type = ["cdylib"]
 [dependencies]
 bytes = "1.8.0"
 http-body-util = "0.1.2"
-hyper = { version = "1.5.1", features = ["client", "http1"] }
-hyper-rustls = { version = "0.27.3", default-features = false, features = ["http1", "native-tokio", "ring", "tls12", "webpki-roots"] }
-hyper-util = { version = "0.1.10", features = ["client", "client-legacy", "http1", "tokio"] }
-pyo3 = { version = "0.28.3", features = ["extension-module"] }
-tokio = { version = "1.41.1", features = ["rt-multi-thread", "time"] }
 url = "2.5.4"
+
+[dependencies.hyper]
+version = "1.5.1"
+features = [
+    "client",
+    "http1",
+]
+
+[dependencies.hyper-rustls]
+version = "0.27.3"
+default-features = false
+features = [
+    "http1",
+    "native-tokio",
+    "ring",
+    "tls12",
+    "webpki-roots",
+]
+
+[dependencies.hyper-util]
+version = "0.1.10"
+features = [
+    "client",
+    "client-legacy",
+    "http1",
+    "tokio",
+]
+
+[dependencies.pyo3]
+version = "0.28.3"
+features = [
+    "extension-module",
+]
+
+[dependencies.tokio]
+version = "1.41.1"
+features = [
+    "rt-multi-thread",
+    "time",
+]
