+++ bench/clients/__init__.py
@@ -1,0 +1,8 @@
+__all__ = (
+    "AsyncClientAdapter",
+    "SyncClientAdapter",
+    "available_clients",
+)
+
+from bench.clients.base import AsyncClientAdapter, SyncClientAdapter
+from bench.clients.registry import available_clients
