--- bench/results/20260507-002101.md
@@ -1,49 +1,0 @@
-# FogHTTP benchmark results
-
-- Timestamp: `20260507-002101`
-- Python: `3.14.0`
-- Platform: `macOS-26.3.1-arm64-arm-64bit-Mach-O`
-- Requests/run: `1000`
-- Warmup/run: `100`
-- Repeats: `2`
-
-## Aggregate
-
-| client | scenario | conc | req/s median | p50 ms | p95 ms | p99 ms | max RSS MB | max threads | max fds | errors |
-| --- | --- | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: |
-| foghttp | bytes-64k | 1 | 5497.7 | 0.16 | 0.25 | 0.40 | 50.6 | 18 | 12 | 0 |
-| foghttp | bytes-64k | 10 | 10928.2 | 0.74 | 1.27 | 1.99 | 62.8 | 27 | 30 | 0 |
-| foghttp | bytes-64k | 50 | 15842.8 | 2.90 | 3.43 | 3.92 | 63.8 | 49 | 42 | 0 |
-| foghttp | bytes-64k | 100 | 13929.2 | 5.69 | 9.42 | 10.09 | 71.1 | 49 | 42 | 0 |
-| foghttp | json-small | 1 | 7412.9 | 0.11 | 0.18 | 0.27 | 42.0 | 3 | 12 | 0 |
-| foghttp | json-small | 10 | 10873.9 | 0.68 | 1.24 | 1.69 | 45.9 | 24 | 30 | 0 |
-| foghttp | json-small | 50 | 15760.2 | 2.52 | 2.99 | 4.14 | 47.9 | 49 | 42 | 0 |
-| foghttp | json-small | 100 | 15753.8 | 4.91 | 5.30 | 5.44 | 49.5 | 49 | 42 | 0 |
-| foghttp | post-echo | 1 | 7660.0 | 0.11 | 0.16 | 0.19 | 77.9 | 18 | 12 | 0 |
-| foghttp | post-echo | 10 | 12125.1 | 0.68 | 0.92 | 1.03 | 78.2 | 27 | 30 | 0 |
-| foghttp | post-echo | 50 | 15776.2 | 2.53 | 3.24 | 3.83 | 78.5 | 49 | 42 | 0 |
-| foghttp | post-echo | 100 | 15848.8 | 5.16 | 5.71 | 5.93 | 78.6 | 49 | 42 | 0 |
-| httpx | bytes-64k | 1 | 1677.6 | 0.52 | 0.73 | 1.01 | 61.5 | 17 | 9 | 0 |
-| httpx | bytes-64k | 10 | 2007.5 | 4.57 | 5.26 | 13.35 | 62.6 | 17 | 27 | 0 |
-| httpx | bytes-64k | 50 | 199.8 | 145.01 | 817.09 | 1203.50 | 70.6 | 17 | 107 | 0 |
-| httpx | bytes-64k | 100 | 169.1 | 184.31 | 2541.10 | 3821.47 | 77.9 | 17 | 207 | 0 |
-| httpx | json-small | 1 | 1864.1 | 0.49 | 0.74 | 1.04 | 48.5 | 2 | 9 | 0 |
-| httpx | json-small | 10 | 1937.9 | 4.33 | 5.06 | 12.00 | 46.6 | 14 | 27 | 0 |
-| httpx | json-small | 50 | 206.1 | 165.14 | 726.69 | 1062.82 | 48.9 | 17 | 107 | 0 |
-| httpx | json-small | 100 | 170.5 | 188.47 | 2443.15 | 3822.59 | 50.6 | 17 | 207 | 0 |
-| httpx | post-echo | 1 | 2055.6 | 0.45 | 0.60 | 0.76 | 77.9 | 17 | 9 | 0 |
-| httpx | post-echo | 10 | 1821.7 | 4.76 | 5.70 | 14.05 | 78.0 | 17 | 27 | 0 |
-| httpx | post-echo | 50 | 204.0 | 160.56 | 734.94 | 1083.01 | 78.0 | 17 | 107 | 0 |
-| httpx | post-echo | 100 | 168.7 | 213.04 | 2332.03 | 3675.23 | 78.0 | 17 | 207 | 0 |
-| zapros | bytes-64k | 1 | 4108.6 | 0.21 | 0.29 | 0.33 | 61.7 | 17 | 9 | 0 |
-| zapros | bytes-64k | 10 | 5575.3 | 1.63 | 1.97 | 2.38 | 62.6 | 17 | 27 | 0 |
-| zapros | bytes-64k | 50 | 5362.8 | 8.37 | 8.71 | 8.83 | 69.5 | 17 | 27 | 0 |
-| zapros | bytes-64k | 100 | 5659.5 | 16.58 | 17.17 | 17.47 | 77.9 | 17 | 27 | 0 |
-| zapros | json-small | 1 | 5216.9 | 0.17 | 0.24 | 0.29 | 44.3 | 2 | 9 | 0 |
-| zapros | json-small | 10 | 6805.1 | 1.25 | 1.52 | 1.99 | 46.7 | 14 | 27 | 0 |
-| zapros | json-small | 50 | 6877.2 | 6.43 | 6.87 | 6.99 | 48.9 | 17 | 27 | 0 |
-| zapros | json-small | 100 | 6124.6 | 14.09 | 18.06 | 18.35 | 50.6 | 17 | 27 | 0 |
-| zapros | post-echo | 1 | 4934.6 | 0.18 | 0.26 | 0.44 | 77.9 | 17 | 9 | 0 |
-| zapros | post-echo | 10 | 6847.1 | 1.34 | 1.59 | 1.91 | 78.0 | 17 | 27 | 0 |
-| zapros | post-echo | 50 | 6444.0 | 6.87 | 8.73 | 9.53 | 78.0 | 17 | 27 | 0 |
-| zapros | post-echo | 100 | 6864.2 | 14.02 | 15.05 | 15.21 | 67.4 | 17 | 27 | 0 |
