--- bench/results/20260507-002126.md
@@ -1,22 +1,0 @@
-# FogHTTP benchmark results
-
-- Timestamp: `20260507-002126`
-- Python: `3.14.0`
-- Platform: `macOS-26.3.1-arm64-arm-64bit-Mach-O`
-- Requests/run: `300`
-- Warmup/run: `50`
-- Repeats: `2`
-
-## Aggregate
-
-| client | scenario | conc | req/s median | p50 ms | p95 ms | p99 ms | max RSS MB | max threads | max fds | errors |
-| --- | --- | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: |
-| foghttp | delay-20ms | 10 | 447.8 | 21.37 | 23.28 | 24.51 | 43.2 | 24 | 30 | 0 |
-| foghttp | delay-20ms | 50 | 694.7 | 65.44 | 85.09 | 87.91 | 45.4 | 49 | 42 | 0 |
-| foghttp | delay-20ms | 100 | 676.7 | 135.49 | 158.74 | 159.50 | 47.5 | 49 | 42 | 0 |
-| httpx | delay-20ms | 10 | 419.5 | 22.50 | 28.99 | 34.68 | 49.6 | 14 | 27 | 0 |
-| httpx | delay-20ms | 50 | 139.7 | 145.63 | 1176.65 | 1803.67 | 46.9 | 17 | 107 | 0 |
-| httpx | delay-20ms | 100 | 127.9 | 545.95 | 1893.37 | 2271.44 | 47.3 | 17 | 107 | 0 |
-| zapros | delay-20ms | 10 | 422.5 | 23.02 | 25.89 | 26.79 | 44.8 | 14 | 27 | 0 |
-| zapros | delay-20ms | 50 | 395.6 | 125.09 | 129.39 | 129.84 | 46.9 | 17 | 27 | 0 |
-| zapros | delay-20ms | 100 | 427.6 | 225.43 | 237.92 | 238.82 | 47.3 | 17 | 27 | 0 |
