--- bench/results/20260507-002954.json
@@ -1,211 +1,0 @@
-{
-  "aggregate": [
-    {
-      "client": "foghttp",
-      "concurrency": 100,
-      "errors_total": 0,
-      "fds_max": 42,
-      "p50_ms_median": 6.822062,
-      "p95_ms_median": 7.7213955,
-      "p99_ms_median": 8.003354,
-      "repeats": 2,
-      "req_s_median": 11286.322577674762,
-      "requests": 1000,
-      "rss_mb_max": 55.140625,
-      "scenario": "post-echo-64k",
-      "threads_max": 49
-    },
-    {
-      "client": "httpx",
-      "concurrency": 100,
-      "errors_total": 0,
-      "fds_max": 207,
-      "p50_ms_median": 243.78733300000002,
-      "p95_ms_median": 2486.3917085000003,
-      "p99_ms_median": 3702.850687,
-      "repeats": 2,
-      "req_s_median": 161.0617131957968,
-      "requests": 1000,
-      "rss_mb_max": 82.203125,
-      "scenario": "post-echo-64k",
-      "threads_max": 17
-    },
-    {
-      "client": "zapros",
-      "concurrency": 100,
-      "errors_total": 0,
-      "fds_max": 27,
-      "p50_ms_median": 18.563708499999997,
-      "p95_ms_median": 25.790458,
-      "p99_ms_median": 26.3112085,
-      "repeats": 2,
-      "req_s_median": 5014.665274469543,
-      "requests": 1000,
-      "rss_mb_max": 81.296875,
-      "scenario": "post-echo-64k",
-      "threads_max": 17
-    }
-  ],
-  "metadata": {
-    "args": {
-      "clients": "foghttp,httpx,zapros",
-      "concurrency": "100",
-      "repeats": 2,
-      "requests": 1000,
-      "scenarios": "post-echo-64k",
-      "warmup": 100
-    },
-    "platform": "macOS-26.3.1-arm64-arm-64bit-Mach-O",
-    "python": "3.14.0 (main, Oct 28 2025, 12:03:45) [Clang 20.1.4 ]",
-    "skipped": {},
-    "timestamp": "20260507-002954"
-  },
-  "runs": [
-    {
-      "client": "foghttp",
-      "client_stats": {
-        "active_connections": 0,
-        "closed_connections": 0,
-        "failed_requests": 0,
-        "idle_connections": 0,
-        "opened_connections": 0,
-        "pending_acquires": 0,
-        "pool_timeouts": 0,
-        "reused_connections": 0,
-        "total_requests": 1100
-      },
-      "concurrency": 100,
-      "duration_s": 0.0917531669838354,
-      "errors": 0,
-      "max_ms": 8.342417,
-      "min_ms": 1.813875,
-      "p50_ms": 6.656458,
-      "p90_ms": 7.581292,
-      "p95_ms": 7.939333,
-      "p99_ms": 8.271417,
-      "peak_fds": 42,
-      "peak_rss_mb": 54.84375,
-      "peak_threads": 49,
-      "process_cpu_s": 0.14811500000000002,
-      "repeat": 1,
-      "requests": 1000,
-      "requests_per_second": 10898.806361377965,
-      "scenario": "post-echo-64k"
-    },
-    {
-      "client": "foghttp",
-      "client_stats": {
-        "active_connections": 0,
-        "closed_connections": 0,
-        "failed_requests": 0,
-        "idle_connections": 0,
-        "opened_connections": 0,
-        "pending_acquires": 0,
-        "pool_timeouts": 0,
-        "reused_connections": 0,
-        "total_requests": 1100
-      },
-      "concurrency": 100,
-      "duration_s": 0.08566162490751594,
-      "errors": 0,
-      "max_ms": 8.438583,
-      "min_ms": 3.286417,
-      "p50_ms": 6.987666,
-      "p90_ms": 7.41525,
-      "p95_ms": 7.503458,
-      "p99_ms": 7.735291,
-      "peak_fds": 42,
-      "peak_rss_mb": 55.140625,
-      "peak_threads": 49,
-      "process_cpu_s": 0.143525,
-      "repeat": 2,
-      "requests": 1000,
-      "requests_per_second": 11673.83879397156,
-      "scenario": "post-echo-64k"
-    },
-    {
-      "client": "httpx",
-      "client_stats": null,
-      "concurrency": 100,
-      "duration_s": 6.435578375007026,
-      "errors": 0,
-      "max_ms": 5285.823708,
-      "min_ms": 17.402041,
-      "p50_ms": 264.913625,
-      "p90_ms": 1681.74425,
-      "p95_ms": 2609.532417,
-      "p99_ms": 3764.216958,
-      "peak_fds": 207,
-      "peak_rss_mb": 82.203125,
-      "peak_threads": 17,
-      "process_cpu_s": 4.047387,
-      "repeat": 1,
-      "requests": 1000,
-      "requests_per_second": 155.38618935689806,
-      "scenario": "post-echo-64k"
-    },
-    {
-      "client": "httpx",
-      "client_stats": null,
-      "concurrency": 100,
-      "duration_s": 5.997460542013869,
-      "errors": 0,
-      "max_ms": 5000.339583,
-      "min_ms": 14.539042,
-      "p50_ms": 222.661041,
-      "p90_ms": 1685.808334,
-      "p95_ms": 2363.251,
-      "p99_ms": 3641.484416,
-      "peak_fds": 207,
-      "peak_rss_mb": 80.21875,
-      "peak_threads": 17,
-      "process_cpu_s": 3.8962200000000005,
-      "repeat": 2,
-      "requests": 1000,
-      "requests_per_second": 166.73723703469554,
-      "scenario": "post-echo-64k"
-    },
-    {
-      "client": "zapros",
-      "client_stats": null,
-      "concurrency": 100,
-      "duration_s": 0.21097758400719613,
-      "errors": 0,
-      "max_ms": 32.974917,
-      "min_ms": 5.702666,
-      "p50_ms": 18.654334,
-      "p90_ms": 25.337167,
-      "p95_ms": 31.803625,
-      "p99_ms": 32.6715,
-      "peak_fds": 27,
-      "peak_rss_mb": 81.296875,
-      "peak_threads": 17,
-      "process_cpu_s": 0.19693200000000033,
-      "repeat": 1,
-      "requests": 1000,
-      "requests_per_second": 4739.840038958317,
-      "scenario": "post-echo-64k"
-    },
-    {
-      "client": "zapros",
-      "client_stats": null,
-      "concurrency": 100,
-      "duration_s": 0.1890541249886155,
-      "errors": 0,
-      "max_ms": 20.028959,
-      "min_ms": 5.595167,
-      "p50_ms": 18.473083,
-      "p90_ms": 19.536666,
-      "p95_ms": 19.777291,
-      "p99_ms": 19.950917,
-      "peak_fds": 27,
-      "peak_rss_mb": 81.296875,
-      "peak_threads": 17,
-      "process_cpu_s": 0.18630500000000083,
-      "repeat": 2,
-      "requests": 1000,
-      "requests_per_second": 5289.490509980769,
-      "scenario": "post-echo-64k"
-    }
-  ]
-}
