+++ bench/templates/report.md.j2
@@ -1,0 +1,26 @@
+# FogHTTP benchmark results
+
+- Timestamp: `{{ timestamp }}`
+- Python: `{{ python_version }}`
+- Platform: `{{ platform_name }}`
+- Server: `local asyncio HTTP/1.1 loopback server`
+- Requests/run: `{{ args.requests }}`
+- Warmup/run: `{{ args.warmup }}`
+- Repeats: `{{ args.repeats }}`
+- Shuffle seed: `{{ args.seed }}`
+
+{% if skipped %}
+## Skipped clients
+
+{% for name, reason in skipped.items() %}
+- `{{ name }}`: {{ reason }}
+{% endfor %}
+
+{% endif %}
+## Aggregate
+
+| mode | client | scenario | conc | conns | ok/s median | req/s median | cv % | p50 ms | p95 ms | p99 ms | errors % | max RSS MB | max threads | max fds |
+| --- | --- | --- | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: | ---: |
+{% for row in aggregate %}
+| {{ row.mode }} | {{ row.client }} | {{ row.scenario }} | {{ row.concurrency }} | {{ row.max_connections }} | {{ "%.1f"|format(row.ok_req_s_median) }} | {{ "%.1f"|format(row.req_s_median) }} | {{ "%.1f"|format(row.req_s_cv_percent) }} | {{ "%.2f"|format(row.p50_ms_median) }} | {{ "%.2f"|format(row.p95_ms_median) }} | {{ "%.2f"|format(row.p99_ms_median) }} | {{ "%.2f"|format(row.error_rate_percent) }} | {{ "%.1f"|format(row.rss_mb_max) }} | {{ row.threads_max }} | {{ row.fds_max }} |
+{% endfor %}
