+++ pyproject.toml
@@ -28,8 +28,21 @@ Repository = "https://github.com/AmberFog/foghttp"
 Issues = "https://github.com/AmberFog/foghttp/issues"
 
 [project.optional-dependencies]
-dev = ["coverage[toml]>=7,<8", "faker>=33", "pytest>=8", "pytest-asyncio>=0.23"]
-bench = ["faker>=33", "httpx>=0.28,<1", "psutil>=6,<8", "zapros>=0.11,<1"]
+dev = [
+    "coverage[toml]>=7,<8",
+    "faker>=33",
+    "pytest>=8",
+    "pytest-asyncio>=0.23",
+]
+bench = [
+    "aiohttp>=3.12,<4",
+    "faker>=33",
+    "httpx>=0.28,<1",
+    "jinja2>=3.1,<4",
+    "psutil>=6,<8",
+    "typer>=0.16,<1",
+    "zapros>=0.11,<1",
+]
 
 [tool.maturin]
 features = ["pyo3/extension-module"]
